DFR-69: OCS: Get JMS msg when DuraStore items are added/deleted
Implements an initial DuraCloud JMS-based listener for use with the OCS,
with unit and integration tests.

This commit also sets up an initial scheme for running integration tests
for this project only when requested by specifying -DskipITs=false. See
the README for more information about how to run integration tests and other
common tasks, including setting up your settings.xml file with the dfrtest
DuraCloud instance password.

This commit also adds an id to the StorageObjectEvent class as a way to
correlate an initial event with the resulting content in Fedora. The DuraCloud
impl populates this with the JMS message id.

https://jira.duraspace.org/browse/DFR-69